Implement title-based slug matching in Collect Templates data job
When collecting template data, some templates use Explorer URLs instead
of direct Grapher links. In these cases, the slug cannot be extracted from
the URL. This change implements a fallback that derives the slug from the
template title (e.g., Template:OWID/Liberal political institutions row ->
liberal-political-institutions-row).

Changes:
- Added `slugify_title` helper in `collect_main_files_worker.py`.
- Updated worker logic to use this helper as a fallback for slug metadata.
- Added unit tests to verify the slugification logic.
