+++ src/main_app/db/sql_schema_tables.py
@@ -124,8 +124,7 @@ owid_charts_templates = """
         t.title AS template_title
     FROM owid_charts c
     LEFT JOIN templates t
-        ON t.slug = c.slug
-        OR (t.slug = '' AND LOWER(REPLACE(REPLACE(REPLACE(t.title, 'Template:OWID/', ''), ' ', '-'), '_', '-')) = c.slug);
+        ON t.slug = c.slug;
 """
 
 templates_need_update = """
@@ -139,7 +138,6 @@ templates_need_update = """
     FROM owid_charts c
     JOIN templates t
         ON t.slug = c.slug
-        OR (t.slug = '' AND LOWER(REPLACE(REPLACE(REPLACE(t.title, 'Template:OWID/', ''), ' ', '-'), '_', '-')) = c.slug)
     WHERE t.last_world_year != c.max_time
     AND t.last_world_year IS NOT NULL
 """
